Split up helper files Split out the subject set data fetching into `helpers/fetchSubjectSets`.
