+++ packages/app-project/src/helpers/fetchSubjectSets/index.js
@@ -1,0 +1,1 @@
+export { default } from './fetchSubjectSets'
